Just changed a couple of array-index references in mutex functions.

+++ my_pthread.c
@@ -302,7 +302,7 @@ int my_pthread_mutex_lock(my_pthread_mutex_t *mutex) {
 			ptr->next = new;
 		}
 		//set thread status to BLOCKED and change context
-		tcbList[current_thread]->status = THREAD_BLOCKED;
+		tcbList[(unsigned int) current_thread]->status = THREAD_BLOCKED;
 		//let the manager continue in the run queue
 		current_thread = MAX_NUM_THREADS + 1;
 		setcontext(&Manager);
@@ -338,7 +338,7 @@ int my_pthread_mutex_unlock(my_pthread_mutex_t *mutex) {
 	pnode *ptr = mutex->waitQueue;
 	mutex->waitQueue = mutex->waitQueue->next;
 	//make this thread ready so it can now acquire this lock
-	tcbList[ptr->tid]->status = THREAD_READY;
+	tcbList[(unsigned int) ptr->tid]->status = THREAD_READY;
 	free(ptr);
 	
 	return 0;
